requirements.txt
flask==3.0.2
joblib==1.3.2
numpy==1.26.4
flask-cors==4.0.0
scikit-learn==1.6.1
gunicorn==21.2.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80 